# Re-implement the 'ExceptT' layer of the parser and switch to megaparsec 7 [WIP]