… hledger-lib/Hledger/Data/Journal.hs
… hledger-lib/Hledger/Reports/ReportOptions.hs
… hledger-lib/Hledger/Utils/Debug.hs
… hledger-lib/Hledger/Utils/Parse.hs
… hledger-lib/Hledger/Utils/Test.hs
… hledger-lib/Text/Megaparsec/Custom.hs
… hledger-lib/hledger-lib.cabal
… hledger-lib/package.yaml
… hledger-ui/Hledger/UI/ErrorScreen.hs
… hledger-ui/hledger-ui.cabal
… hledger-ui/package.yaml
… hledger-web/Hledger/Web/Widget/AddForm.hs
… hledger-web/hledger-web.cabal
… hledger-web/package.yaml
… hledger/Hledger/Cli/Commands/Add.hs
… hledger/Hledger/Cli/Commands/Rewrite.hs
… hledger/hledger.cabal
… hledger/package.yaml
… stack-ghc7.10.yaml
… stack-ghc8.0.yaml
… stack-ghc8.2.yaml
… stack.yaml
… tests/journal/parse-errors.test