Details can be found in the commit messages.

This PR also fixes an issue found in #853.
Error messages for cyclic include files now appear as follows:
```
$ cat a.j 
include b.j

$ cat b.j 
include c.j

$ cat c.j 
include a.j

$ hledger stats -f a.j
hledger: in file included from a.j,
in file included from b.j,
c.j:1:12:
  |
1 | include a.j
  |            ^
Cyclic include: a.j

$ hledger stats -f b.j
hledger: in file included from b.j,
in file included from c.j,
a.j:1:12:
  |
1 | include b.j
  |            ^
Cyclic include: b.j

$ hledger stats -f c.j
hledger: in file included from c.j,
in file included from a.j,
b.j:1:12:
  |
1 | include c.j
  |            ^
Cyclic include: c.j

```