@awjchen 2018-10-01
I've added a commit switching hledger to megaparsec 7, addressing #879.

I ran into a problem while trying to build on ghc 7.10 (we must use config-ini-0.2.3.0 for megaparsec 7, but it uses `-fno-warn-redundant-constraints`, which was only introduced in ghc 8.0) that I don't know how to resolve, so I've marked this PR as a WIP. Any suggestions would be greatly appreciated!

Also, benchmarks with `stack bench hledger` do not indicate a regression:

Before
```
read bench/10000x1000x10.journal        [1.27s]
print                                   [1.61s]
register                                [2.26s]
balance                                 [0.18s]
Total: 5.31s                 
```
After adding the `ExceptT` layer
```
read bench/10000x1000x10.journal        [1.14s]
print                                   [1.59s]
register                                [2.22s]
balance                                 [0.18s]
Total: 5.13s                 
```
After switching to megaparsec 7
```
read bench/10000x1000x10.journal        [1.03s]
print                                   [1.60s]
register                                [2.28s]
balance                                 [0.20s]
Total: 5.11s                 
```