@simonmichael 2018-10-05
Thanks for this, and for checking performance.

It's fine to ignore GHC 7.10 at this point. We can remove it officially from the supported GHCs later.

Here's my take on some related issues:

|          |              |                                                                                                          |                |
| -------- | ------------ | -------------------------------------------------------------------------------------------------------- | -------------- |
| issue/pr | author       | title/description                                                                                        | waiting for    |
| \#807    | simonmichael | allow descriptions/comments in periodic transactions (regression in periodic transaction parsing)        | \#880          |
| \#823    | awjchen      | Allow "re-parsing" with custom parse errors, some commodity cleanups (& fix 807)                         | ?              |
| \#853    |              | Hledger does not detect cycles when resolving 'include' directives                                       | \#880          |
| \#879    |              | Plan for megaparsec 7 support? blocking Arch Linux packagers                                             | \#880          |
| \#880    | awjchen      | Re-implement the 'ExceptT' layer of the parser and switch to megaparsec 7 \[WIP\] (fixing 807? 823? 853) | review, rebase |