@simonmichael 2018-10-05
> Hmm, I think I can do better with the comments so that the code is less opaque. I will do some work on them, but it might be one or two days before I have the chance to. Is there anything particularly opaque that I should make sure to address?

Nothing particular, unless we can think of a better name than "Bundle". After some more time/experience we might find more ways to clarify.