@simonmichael 2018-10-05
I confirm this fixes the wrong parsing error we were triggering on #853. 

I thought this might be related to #823 as well, but I'm not sure. Anyway I guess I'll wait till this lands before looking further at #807/#823.