@awjchen 2018-10-09
I've added some commits expanding the comments and updating the megaparsec bounds.

In this PR I was focused only on #853 and #879. I haven't thought of #807 or #823 in a while, so I'm not yet sure how to proceed with them, but I think that work should be done next.