@simonmichael 2018-10-09
> I ran into a problem while trying to build on ghc 7.10 (we must use config-ini-0.2.3.0 for megaparsec 7, but it uses -fno-warn-redundant-constraints, which was only introduced in ghc 8.0)

Reported: https://github.com/aisamanra/config-ini/issues/1#issuecomment-428377098