@simonmichael 2018-10-24
Fixed by #916.